Fix "cpm_add_patches" handling of multiple patch commands
Update "cpm_add_patches" to correctly handle multiple patch
files. The "PATCH_COMMAND" executor may not be a shell, so
it is not portable to use shell operators like "&&" and "<"
in the command. Commands after the initial PATCH_COMMAND
should nstead be prefixed by COMMAND, and we can use "-i"
to specify that a file is used by the "patch" command.